app/requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
torch==2.1.0+cpu
numpy==1.26.4
--extra-index-url https://download.pytorch.org/whl/cpu